This is the zoom out button and standardization of the timing data
Hi Dr. Zhu, like the commit says, this is the zoom out button and timing standardization for the animation. The zoom out is a little weird, just cause it was hard to have the figure carry through between each call back function (the one for the click event, and the button click). It will zoom out to the furthest range though.